homeTask/ItStep/lesson16/telegram_bot/requirements.txt
psutil~=7.0.0
python-telegram-bot~=22.1
python-dotenv~=1.1.0